+++ src/mesido/heat_physics_mixin.py
@@ -389,14 +389,22 @@ class HeatPhysicsMixin(
             if len(temperature_regimes) == 0:
                 temperature = temperatures["temperature"]
                 self.__temperature_regime_var_bounds[temp_var_name] = (temperature, temperature)
-            elif len(temperature_regimes) == 1:
-                temperature = temperature_regimes[0]
-                self.__temperature_regime_var_bounds[temp_var_name] = (temperature, temperature)
             else:
-                self.__temperature_regime_var_bounds[temp_var_name] = (
-                    min(temperature_regimes),
-                    max(temperature_regimes),
-                )
+                if max(temperature_regimes) >= temperatures["temperature"]:
+                    logger.error(
+                        f"The temperature provided for carrier with name "
+                        f"{temperatures['name']} and id {temperatures['id']} is "
+                        f"smaller than the largest value in the temperature regime "
+                        f"provided for this carrier, please update the esdl."
+                    )
+                if len(temperature_regimes) == 1:
+                    temperature = temperature_regimes[0]
+                    self.__temperature_regime_var_bounds[temp_var_name] = (temperature, temperature)
+                else:
+                    self.__temperature_regime_var_bounds[temp_var_name] = (
+                        min(temperature_regimes),
+                        max(temperature_regimes),
+                    )
 
             for temperature_regime in temperature_regimes:
                 carrier_selected_var = carrier_id_number_mapping + f"_{temperature_regime}"
@@ -1398,6 +1406,7 @@ class HeatPhysicsMixin(
 
             sup_carrier = parameters[f"{s}.T_supply_id"]
             supply_temperatures = self.temperature_regimes(sup_carrier)
+
             big_m = 2.0 * self.bounds()[f"{s}.HeatOut.Heat"][1]
             big_m = (
                 big_m
@@ -1405,26 +1414,91 @@ class HeatPhysicsMixin(
                 else 2.0 * self.bounds()[f"{s}.Heat_source"][1] * parameters[f"{s}.T_supply"] / dt
             )
 
-            if len(supply_temperatures) == 0:
+            # Check to see if the out carrier has a temperature profile assigned to it.
+            temp_out_profile, _, _, _ = self.__get_out_port_carrier_temp_profile(
+                parameters, s, "heat_source"
+            )
+
+            if temp_out_profile is None:
+                if len(supply_temperatures) == 0:
+                    constraints.append(
+                        (
+                            (heat_out - discharge * cp * rho * parameters[f"{s}.T_supply"])
+                            / heat_nominal,
+                            0.0,
+                            0.0,
+                        )
+                    )
+                else:
+                    for supply_temperature in supply_temperatures:
+                        sup_temperature_is_selected = self.state(
+                            f"{sup_carrier}_{supply_temperature}"
+                        )
+
+                        constraints.extend(
+                            self._symmetric_big_m_constraints(
+                                heat_out - discharge * cp * rho * supply_temperature,
+                                (1.0 - sup_temperature_is_selected) * big_m,
+                                constraint_nominal,
+                            )
+                        )
+
+        return constraints
+
+    def __source_heat_to_discharge_variable_temp_constraints(self, ensemble_member):
+        """
+        Adds the same type of constraints to the source as
+        __source_heat_to_discharge_path_constraints for cases where the out carrier
+        has a prescribed temperature profile. An important difference is that these
+        are conventional constraints, since every timestep will have a specific value.
+
+        TODO: This has not yet been tested and developed for cases with multiple sources on the
+        same network and only one contains the temperature profile. With the current
+        implementation, all source on the same network should have this same temperature profile.
+        """
+
+        constraints = []
+        parameters = self.parameters(ensemble_member)
+
+        for s in self.energy_system_components.get("heat_source", []):
+            heat_nominal = parameters[f"{s}.Heat_nominal"]
+            cp = parameters[f"{s}.cp"]
+            rho = parameters[f"{s}.rho"]
+            dt = parameters[f"{s}.dT"]
+
+            big_m = 2.0 * self.bounds()[f"{s}.HeatOut.Heat"][1]
+            big_m = (
+                big_m
+                if big_m != np.inf
+                else 2.0 * self.bounds()[f"{s}.Heat_source"][1] * parameters[f"{s}.T_supply"] / dt
+            )
+
+            temp_out_profile, sup_carrier_name, temp_out_prof_start_idx, temp_out_prof_end_idx = (
+                self.__get_out_port_carrier_temp_profile(parameters, s, "heat_source")
+            )
+
+            if (
+                temp_out_profile is not None
+            ):  # Case where the out carrier has a temp profile assigned to it.
+                heat_out_vector = self.__state_vector_scaled(f"{s}.HeatOut.Heat", ensemble_member)
+                discharge_vector = self.__state_vector_scaled(f"{s}.Q", ensemble_member)
+
                 constraints.append(
                     (
-                        (heat_out - discharge * cp * rho * parameters[f"{s}.T_supply"])
+                        (
+                            heat_out_vector
+                            - discharge_vector
+                            * cp
+                            * rho
+                            * temp_out_profile.values[
+                                temp_out_prof_start_idx : temp_out_prof_end_idx + 1
+                            ]
+                        )
                         / heat_nominal,
                         0.0,
                         0.0,
                     )
                 )
-            else:
-                for supply_temperature in supply_temperatures:
-                    sup_temperature_is_selected = self.state(f"{sup_carrier}_{supply_temperature}")
-
-                    constraints.extend(
-                        self._symmetric_big_m_constraints(
-                            heat_out - discharge * cp * rho * supply_temperature,
-                            (1.0 - sup_temperature_is_selected) * big_m,
-                            constraint_nominal,
-                        )
-                    )
 
         return constraints
 
@@ -1477,6 +1551,56 @@ class HeatPhysicsMixin(
                     )
 
         return constraints
+
+    def __get_out_port_carrier_temp_profile(self, parameters, asset_name, asset_type):
+        """
+        This function finds the carrier lined to the asset's out port and grabs the
+        temperature profile assigned to it, if there is one assigned to it.
+        It returns the temperature as a timeseries, the name of the carrier, and the
+        start and end index of the temperature profile according to the problem's timeseries
+        (this last one only relevant for problems that are sliced).
+        """
+        # TODO: modify profile parser so the temperature profile is not called a price profile.
+        # TODO: modify this and the profile parser to incorporate the esdl option to
+        # use power instead of temperature.
+        # TODO: the start/end indices are needed for a very specific problem-times slicing case.
+        # Modify the way problems are sliced to remove this.
+
+        try:
+            carriers = self.esdl_carriers
+            carriers_ids = carriers.keys()
+        except AttributeError:
+            carriers = None
+            carriers_ids = []
+            sup_carrier_name = None
+        temp_out_profile = None
+        temp_out_prof_start_idx = None
+        temp_out_prof_end_idx = None
+        carrier_id_types = {"heat_source": ".T_supply_id", "heat_pipe": ".carrier_id"}
+        for carrier_id in carriers_ids:
+            if (
+                carriers[carrier_id]["id_number_mapping"]
+                == parameters[f"{asset_name}{carrier_id_types[asset_type]}"]
+            ):
+                sup_carrier_name = carriers[carrier_id]["name"]
+        try:
+            temp_out_profile = self.get_timeseries(f"{sup_carrier_name}.price_profile")
+            temp_out_prof_start_idx = int(
+                np.where(
+                    self.get_timeseries(f"{sup_carrier_name}.price_profile").times
+                    == self.times()[0]
+                )[0]
+            )
+            temp_out_prof_end_idx = int(
+                np.where(
+                    self.get_timeseries(f"{sup_carrier_name}.price_profile").times
+                    == self.times()[-1]
+                )[0]
+            )
+        except KeyError:
+            pass
+
+        return temp_out_profile, sup_carrier_name, temp_out_prof_start_idx, temp_out_prof_end_idx
 
     def __pipe_heat_to_discharge_path_constraints(self, ensemble_member):
         """
@@ -1532,7 +1656,13 @@ class HeatPhysicsMixin(
             # negative or possitive based on flowdir
 
             for heat in [scaled_heat_in, scaled_heat_out]:
-                if self.energy_system_options()["neglect_pipe_heat_losses"]:
+                temp_out_profile, _, _, _ = self.__get_out_port_carrier_temp_profile(
+                    parameters, p, "heat_pipe"
+                )
+                if (
+                    self.energy_system_options()["neglect_pipe_heat_losses"]
+                    and temp_out_profile is None
+                ):
                     temp = parameters[f"{p}.temperature"]
                     if len(temperatures) == 0:
                         constraints.append(
@@ -1552,7 +1682,7 @@ class HeatPhysicsMixin(
                                     big_m,
                                 )
                             )
-                else:
+                elif not self.energy_system_options()["neglect_pipe_heat_losses"]:
                     # Note that during cold delivery the line can be colder than the ground
                     # temperature.
                     # In this case we have to bound the heat flowing in the line with the ground
@@ -3484,6 +3614,9 @@ class HeatPhysicsMixin(
             constraints.extend(self.__heat_matching_demand_insulation_constraints(ensemble_member))
 
         constraints.extend(self.__ates_max_stored_heat_constraints(ensemble_member))
+        constraints.extend(
+            self.__source_heat_to_discharge_variable_temp_constraints(ensemble_member)
+        )
 
         return constraints